requirements.txt
feedparser
requests
newspaper3k
lxml_html_clean
faiss-cpu
numpy
fastapi
uvicorn
scikit-learn
transformers
torch
